Refactor authenticaiton handling to separate class
